adding cleaned up classes
