+++ sdkproject/Assets/Mapbox/Unity/Editor/PropertyDrawers/GameplaySectionDrawer.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: cf58941459c8146968b0fb8f5a82de58
+timeCreated: 1529622130
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
