+++ sdkproject/Assets/Mapbox/Unity/Editor/PropertyDrawers/ModelingSectionDrawer.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 39f85c0d2d7cb4b5e989cf0edd7109a9
+timeCreated: 1529620234
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
