+++ sdkproject/Assets/Mapbox/Unity/LayerProperties/PrefabItemOptions.cs
@@ -16,7 +16,7 @@ namespace Mapbox.Unity.Map
 		public readonly VectorPrimitiveType primitiveType = VectorPrimitiveType.Point;
 
 		//Group features turned off
-		public readonly bool groupFeatures = false;
+		public readonly bool combineMeshes = false;
 
 		//No extrusion
 		public readonly ExtrusionType extrusionType = ExtrusionType.None;
