+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/LayerVisualizers/LocationPrefabsLayerVisualizer.cs
@@ -40,7 +40,7 @@ namespace Mapbox.Unity.MeshGeneration.Interfaces
 			};
 
 				
-			item.coreOptions.combineMeshes = item.groupFeatures;
+			item.coreOptions.combineMeshes = item.combineMeshes;
 			item.moveFeaturePositionTo = item._movePrefabFeaturePositionTo;
